Projects/ETL claims process MLOPS/requirements.txt
# PySpark & Data Processing
pyspark==3.5.1
pandas>=2.0.0
numpy>=1.24.0

# Delta Lake for ACID transactions
delta-spark==3.1.0

# Arrow for Parquet/interoperability
pyarrow>=12.0.0

# Data validation
pydantic>=2.0.0

# MLflow for experiment tracking
mlflow>=2.10.0

# Machine Learning
scikit-learn>=1.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Data generation
faker>=15.0.0

# Logging
python-dotenv>=1.0.0
loguru>=0.7.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
